+++ src/config/cli.rs
@@ -33,12 +33,12 @@ pub fn build_cli() -> Command {
     let show_l =
         Arg::new("show").short('s')
                         .long("show")
-                        .value_name("c,m,u,s,a")
+                        .value_name("r,m,u,s,a")
                         .display_order(3)
                         .help_heading("Filter")
-                        .help("Show (c)commands, (m)erges, (u)nmerges, (s)yncs, and/or (a)ll")
+                        .help("Show emerge (r)uns, (m)erges, (u)nmerges, (s)yncs, and/or (a)ll")
                         .long_help("Show (any combination of)\n  \
-                                    c: Emerge command\n  \
+                                    r: Emerge runs\n  \
                                     m: Package merges\n  \
                                     u: Package unmerges\n  \
                                     s: Repository syncs\n  \
@@ -46,24 +46,24 @@ pub fn build_cli() -> Command {
     let show_s =
         Arg::new("show").short('s')
                         .long("show")
-                        .value_name("c,p,t,s,a")
+                        .value_name("r,p,t,s,a")
                         .display_order(3)
                         .help_heading("Filter")
-                        .help("Show (c)commands, (p)ackages, (t)otals, (s)yncs, and/or (a)ll")
+                        .help("Show emerge (r)uns, (p)ackages, (t)otals, (s)yncs, and/or (a)ll")
                         .long_help("Show (any combination of)\n  \
-                                    c: Emerge commands\n  \
+                                    r: Emerge runs\n  \
                                     p: Individual package merges/unmerges\n  \
                                     t: Total package merges/unmerges\n  \
                                     s: Repository syncs\n  \
                                     a: All of the above");
     let show_p = Arg::new("show").short('s')
                                  .long("show")
-                                 .value_name("e,m,t,a")
+                                 .value_name("r,m,t,a")
                                  .display_order(3)
                                  .help_heading("Filter")
-                                 .help("Show (e)emerge processes, (m)erges, (t)otal, and/or (a)ll")
+                                 .help("Show (r)unning processes, (m)erges, (t)otal, and/or (a)ll")
                                  .long_help("Show (any combination of)\n  \
-                                             e: Current emerge processes\n  \
+                                             r: Running emerge processes\n  \
                                              m: Package merges\n  \
                                              t: Total estimate\n  \
                                              a: All of the above");